Cambios a rutas
Se implementaron las rutas en la landingpage
recomendable borrar el proyecto local y clonar nuevamente
