+++ README.md
@@ -15,6 +15,10 @@ npm run serve
 npm run build
 ```
 
+### Lints and fixes files
+```
+npm run lint
+```
+
 ### Customize configuration
-See [Configuration Reference](https://cli.vuejs.org/config/).
- -->
+See [Configuration Reference](https://cli.vuejs.org/config/). -->
