+++ src/components/HelloWorld.vue
@@ -15,7 +15,4 @@ export default {
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
-h3 {
-  margin: 40px 0 0;
-}
 </style>
