+++ src/views/AlbunesView.vue
@@ -1,0 +1,13 @@
+<template>
+    <div class="albunes">
+        <h1>Albunes</h1>
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style lang="scss" scoped>
+
+</style>
