+++ src/views/ConciertosView.vue
@@ -1,0 +1,13 @@
+<template>
+    <div class="concierto">
+        <h1>Conciertos</h1>
+    </div>
+</template>
+
+<script>
+
+</script>
+
+<style scoped>
+
+</style>
